requirements.txt
Django>=4.0
djangorestframework>=3.14.0
django-filter>=23.0
psycopg2>=2.9.0
pytest-django>=4.5.0
djangorestframework-simplejwt>=5.2.0
gunicorn>=21.0
django-simple-history>=3.8.0
python-decouple>=3.8